refactor: admin 도메인 중심 패키지 리팩토링
