+++ src/main/java/net/detalk/api/admin/controller/response/VisitorLogResponse.java
@@ -1,4 +1,4 @@
-package net.detalk.api.controller.v1.response;
+package net.detalk.api.admin.controller.response;
 
 import java.time.LocalDateTime;
